+++ lib/boruta/oauth/schemas/client.ex
@@ -46,10 +46,15 @@ defmodule Boruta.Oauth.Client do
   end
 
   @spec check_secret(client :: t(), secret :: String.t()) :: :ok | {:error, String.t()}
-  def check_secret(%__MODULE__{secret: client_secret}, secret) do
-    case client_secret == secret do
+  def check_secret(%__MODULE__{secret: secret}, secret), do: :ok
+  def check_secret(_client, _secret), do: {:error, "Invalid client secret."}
+
+  @spec check_redirect_uri(client :: t(), redirect_uri :: String.t()) ::
+          :ok | {:error, String.t()}
+  def check_redirect_uri(%__MODULE__{redirect_uris: client_redirect_uris}, redirect_uri) do
+    case Enum.member?(client_redirect_uris, redirect_uri) do
       true -> :ok
-      false -> {:error, "Invalid client secret."}
+      false -> {:error, "Client redirect_uri do not match."}
     end
   end